Cleaned up elements of importing dungeoneerCommon
