+++ dungeoneerClient/src/app/display/dm-search-card/dm-search-card.component.spec.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormGroup } from '@angular/forms';
-import { Schema } from 'dungeoneer-common/dist/types/src/schema/schemaTypes';
+import { Schema } from 'dungeoneer-common';
 import { Subject } from 'rxjs';
 import { DmFormInputData } from 'src/app/form/DmFormInputData';
 import { DmFormUtils } from 'src/app/form/DmFormUtils';
