+++ dungeoneerClient/src/app/dm-character-sheet/dm-abstract-character.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Input } from "@angular/core";
 import { getDungeoneerSchema } from "dungeoneer-common";
-import { Schema } from "dungeoneer-common/dist/types/src/schema/schemaTypes";
+import { Schema } from "dungeoneer-common";
 import { DmUnsubscriberComponent } from "../core/dm-unsubscriber/dm-unsubscriber.component";
 import { DmFormEditService } from "../form/dm-form-edit/dm-form-edit.service";
 import { CreatureStats } from "./model/creature-stats.model";
