+++ dungeoneerCommon/src/schema/schemaTypes.ts
@@ -5,13 +5,11 @@ export interface Schema {
     }
 }
 
-
 // nodetype is equivalent to a table in SQL databases.
 export interface NodeType {
 	nodeVars: {
         [key: string]: NodeVar
     }
-	
 	// label var indicates which value in NodeVars is to be used for labels (in tables etc.)
 	// If this is not set "name" is used (if a name var exists)
 	labelVar?: string
