+++ dungeoneerCommon/tsconfig.json
@@ -1,17 +1,12 @@
 {
-    "compilerOptions": {
-        /* Basic Options */
-        "target": "ESNext",                          /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017','ES2018' or 'ESNEXT'. */
-        "module": "ESNext",                     /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */
-        "sourceMap": true,                     /* Generates corresponding '.map' file. */
-        "outDir": "dist",                        /* Redirect output structure to the directory. */
-        "strict": true,
-        "esModuleInterop": true,
-        "moduleResolution": "Node",
-        "declaration": true,
-        "declarationDir": "dist/types"
-      },
-      "exclude": [
-        "./dist/types/**/*"
-      ]
+  "compilerOptions": {
+    /* Basic Options */
+    "target": "ESNext",                          /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017','ES2018' or 'ESNEXT'. */
+    "module": "ESNext",                     /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */
+    "sourceMap": true,                     /* Generates corresponding '.map' file. */
+    "outDir": "dist",                        /* Redirect output structure to the directory. */
+    "strict": true,
+    "esModuleInterop": true,
+    "moduleResolution": "Node"
   }
+}
