+++ dungeoneerServer/src/setter/dgraphSetterTestData.ts
@@ -1,5 +1,5 @@
-import { DmLinkSetParams, DmSetParams } from "dungeoneer-common/dist/types/src/connection/connectionTypes"
-import { Schema } from "dungeoneer-common/dist/types/src/schema/schemaTypes"
+import { DmLinkSetParams, DmSetParams } from "dungeoneer-common"
+import { Schema } from "dungeoneer-common"
 
 const setterSchema1: Schema = {
 	nodeTypes: {
