+++ dungeoneerServer/types/src/queryGenerator/DmDgraphQueryGenerator.d.ts
@@ -1,3 +1,3 @@
-import { DmFetchParams } from "dungeoneer-common/dist/types/src/connection/connectionTypes";
-import { Schema } from "dungeoneer-common/dist/types/src/schema/schemaTypes";
+import { DmFetchParams } from "dungeoneer-common";
+import { Schema } from "dungeoneer-common";
 export declare function generateQuery(params: DmFetchParams, schema: Schema): string;
