+++ dungeoneerServer/types/src/queryGenerator/QueryGenerator.d.ts
@@ -1,5 +1,5 @@
-import { DmFetchParams } from "dungeoneer-common/dist/types/src/connection/connectionTypes";
-import { Schema } from "dungeoneer-common/dist/types/src/schema/schemaTypes";
+import { DmFetchParams } from "dungeoneer-common";
+import { Schema } from "dungeoneer-common";
 export declare class QueryGenerator {
     constructor();
     generateQuery(params: DmFetchParams, schema: Schema): string;
